+++ tests/test_worker.py
@@ -1,8 +1,6 @@
 import asyncio
 import logging
-import time
 from contextlib import asynccontextmanager
-from contextvars import ContextVar
 from datetime import datetime, timedelta, timezone
 from typing import AsyncGenerator, Callable
 from unittest.mock import AsyncMock, patch
@@ -14,19 +12,17 @@ from redis.asyncio import Redis
 from redis.exceptions import ConnectionError
 
 from docket import (
-    ConcurrencyLimit,
     CurrentDocket,
-    CurrentExecution,
     CurrentWorker,
     Docket,
     Perpetual,
     Worker,
 )
 from docket.dependencies import Timeout
-from tests._key_leak_checker import KeyCountChecker
 from docket.execution import Execution
 from docket.tasks import standard_tasks
 from docket.worker import ms
+from tests._key_leak_checker import KeyCountChecker
 
 
 async def test_worker_acknowledges_messages(
@@ -164,156 +160,6 @@ async def test_worker_handles_redeliveries_from_abandoned_workers(
         await worker_b.run_until_finished()
 
     the_task.assert_awaited_once_with()
-
-
-async def test_redeliveries_respect_concurrency_limits(docket: Docket):
-    """Test that redelivered tasks still respect concurrency limits"""
-    task_executions: list[tuple[int, float]] = []  # (customer_id, timestamp)
-    failure_count = 0
-
-    async def task_that_sometimes_fails(
-        customer_id: int,
-        should_fail: bool,
-        concurrency: ConcurrencyLimit = ConcurrencyLimit(
-            "customer_id",
-            max_concurrent=1,  # Only 1 task per customer at a time
-        ),
-    ):
-        nonlocal failure_count
-
-        # Record when this task runs
-
-        task_executions.append((customer_id, time.time()))
-
-        await asyncio.sleep(0.02)  # Brief work
-
-        if should_fail:
-            failure_count += 1
-            raise ValueError("Intentional failure for testing")
-
-    # Schedule tasks: some will fail initially, others succeed
-    await docket.add(task_that_sometimes_fails)(
-        customer_id=1, should_fail=True
-    )  # Will fail first time
-    await docket.add(task_that_sometimes_fails)(
-        customer_id=1, should_fail=False
-    )  # Should queue after first
-    await docket.add(task_that_sometimes_fails)(
-        customer_id=2, should_fail=False
-    )  # Different customer, can run parallel
-    await docket.add(task_that_sometimes_fails)(
-        customer_id=1, should_fail=False
-    )  # More work for customer 1
-
-    # Use short redelivery timeout so failures get redelivered quickly
-    # Must be long enough for cleanup operations to complete before redelivery
-    async with Worker(
-        docket, concurrency=5, redelivery_timeout=timedelta(milliseconds=200)
-    ) as worker:
-        await worker.run_until_finished()
-
-    # Verify all tasks eventually executed
-    customer_1_executions = [t for t in task_executions if t[0] == 1]
-    customer_2_executions = [t for t in task_executions if t[0] == 2]
-
-    # Should have executed all tasks for both customers
-    assert (
-        len(customer_1_executions) == 3
-    )  # One failed task that retried + 2 successful
-    assert len(customer_2_executions) == 1
-
-    # Verify tasks for customer 1 didn't run concurrently (max_concurrent=1)
-    customer_1_times = sorted([t[1] for t in customer_1_executions])
-    for i in range(len(customer_1_times) - 1):
-        # Each task should finish before the next starts (with some buffer for timing)
-        time_gap = customer_1_times[i + 1] - customer_1_times[i]
-        assert time_gap >= 0.015, (
-            f"Tasks for customer 1 overlapped: gap={time_gap:.3f}s"
-        )
-
-    # Should have had at least one failure that was redelivered
-    assert failure_count >= 1
-
-
-async def test_concurrency_blocked_task_executes_exactly_once(docket: Docket):
-    """Tasks blocked by concurrency and rescheduled should execute exactly once,
-    not be duplicated via both rescheduling and redelivery."""
-
-    execution_count = 0
-    execution_keys: list[str] = []
-
-    async def tracked_task(
-        customer_id: int,
-        execution: Execution = CurrentExecution(),
-        concurrency: ConcurrencyLimit = ConcurrencyLimit(
-            "customer_id",
-            max_concurrent=1,
-        ),
-    ) -> None:
-        nonlocal execution_count
-        execution_keys.append(execution.key)
-        execution_count += 1
-        await asyncio.sleep(0.02)
-
-    # Schedule 3 tasks for the same customer - all will be blocked by concurrency
-    await docket.add(tracked_task)(customer_id=1)
-    await docket.add(tracked_task)(customer_id=1)
-    await docket.add(tracked_task)(customer_id=1)
-
-    # Use short redelivery timeout to make race condition more likely
-    async with Worker(
-        docket, concurrency=5, redelivery_timeout=timedelta(milliseconds=50)
-    ) as worker:
-        await worker.run_until_finished()
-
-    # Each task should execute exactly once
-    assert execution_count == 3, (
-        f"Expected 3 executions, got {execution_count}. "
-        f"Execution keys: {execution_keys}"
-    )
-
-    # All execution keys should be unique (no duplicate executions)
-    assert len(set(execution_keys)) == 3, (
-        f"Found duplicate executions: {execution_keys}"
-    )
-
-    # Verify no messages left pending (all were acknowledged)
-    async with docket.redis() as redis:
-        pending_info = await redis.xpending(
-            name=docket.stream_key,
-            groupname=docket.worker_group_name,
-        )
-        assert pending_info["pending"] == 0, (
-            "Found unacknowledged messages - atomic reschedule failed"
-        )
-
-        # Verify stream is empty (cleanup happened)
-        assert await redis.xlen(docket.stream_key) == 0
-
-
-async def test_concurrency_limited_task_successfully_acquires_slot(docket: Docket):
-    """Tasks with concurrency limits successfully acquire slots when available"""
-
-    executed: list[int] = []
-
-    async def limited_task(
-        customer_id: int,
-        concurrency: ConcurrencyLimit = ConcurrencyLimit(
-            "customer_id",
-            max_concurrent=2,
-        ),
-    ) -> None:
-        executed.append(customer_id)
-        await asyncio.sleep(0.01)
-
-    # Schedule a task that has concurrency limits but should acquire slot successfully
-    await docket.add(limited_task)(customer_id=1)
-
-    async with Worker(docket, concurrency=5) as worker:
-        await worker.run_until_finished()
-
-    # Task should have executed successfully
-    assert executed == [1]
 
 
 async def test_worker_handles_unregistered_task_execution_on_initial_delivery(
@@ -658,621 +504,6 @@ async def test_worker_can_be_told_to_skip_automatic_tasks(docket: Docket):
     assert not called
 
 
-async def test_worker_concurrency_limits_task_queuing_behavior(docket: Docket):
-    """Test that concurrency limits control task execution properly"""
-
-    # Use contextvar for reliable tracking across async execution
-    execution_log: ContextVar[list[tuple[str, int]]] = ContextVar("execution_log")
-    execution_log.set([])
-
-    async def task_with_concurrency(
-        customer_id: int,
-        concurrency: ConcurrencyLimit = ConcurrencyLimit(
-            "customer_id", max_concurrent=2
-        ),
-    ):
-        # Record execution start
-        log = execution_log.get()
-        log.append(("start", customer_id))
-        execution_log.set(log)
-
-        # Small delay to ensure concurrency is tested
-        await asyncio.sleep(0.01)
-
-        # Record execution end
-        log = execution_log.get()
-        log.append(("end", customer_id))
-        execution_log.set(log)
-
-    # Schedule tasks for both customers to ensure all branches execute
-    await docket.add(task_with_concurrency)(customer_id=1)
-    await docket.add(task_with_concurrency)(customer_id=2)
-    await docket.add(task_with_concurrency)(customer_id=1)
-    await docket.add(task_with_concurrency)(customer_id=2)
-    await docket.add(task_with_concurrency)(customer_id=1)
-
-    # Run tasks with reasonable concurrency
-    async with Worker(docket, concurrency=5) as worker:
-        await worker.run_until_finished()
-
-    # Verify execution happened
-    log = execution_log.get()
-    start_events = [event for event in log if event[0] == "start"]
-    end_events = [event for event in log if event[0] == "end"]
-
-    # Verify both customer types executed
-    customer_1_starts = len([e for e in start_events if e[1] == 1])
-    customer_2_starts = len([e for e in start_events if e[1] == 2])
-
-    assert customer_1_starts == 3, (
-        f"Expected 3 customer_id=1 tasks, got {customer_1_starts}"
-    )
-    assert customer_2_starts == 2, (
-        f"Expected 2 customer_id=2 tasks, got {customer_2_starts}"
-    )
-    assert len(start_events) == len(end_events) == 5
-
-
-async def test_worker_concurrency_missing_argument_bypass(docket: Docket):
-    """Test that tasks with missing concurrency arguments bypass concurrency control"""
-    task_executed = False
-
-    async def task_missing_concurrency_arg(
-        customer_id: int,
-        concurrency: ConcurrencyLimit = ConcurrencyLimit(
-            "missing_param",
-            max_concurrent=1,  # "missing_param" is not provided
-        ),
-    ):
-        nonlocal task_executed
-        task_executed = True
-
-    # Schedule task without the required parameter - should trigger bypass path
-    await docket.add(task_missing_concurrency_arg)(customer_id=1)
-
-    async with Worker(docket) as worker:
-        await worker.run_until_finished()
-
-    # Task should execute despite missing concurrency argument (bypass path)
-    assert task_executed
-
-
-async def test_worker_concurrency_no_limit_early_return(docket: Docket):
-    """Test tasks without concurrency limits execute normally"""
-    task_executed = False
-
-    async def task_without_concurrency(customer_id: int):
-        nonlocal task_executed
-        task_executed = True
-
-    await docket.add(task_without_concurrency)(customer_id=1)
-
-    async with Worker(docket) as worker:
-        await worker.run_until_finished()
-
-    assert task_executed
-
-
-async def test_worker_concurrency_different_customer_branches(docket: Docket):
-    """Test that different customer IDs are handled in separate branches"""
-    customers_executed: set[int] = set()
-
-    async def track_customer_execution(
-        customer_id: int,
-        concurrency: ConcurrencyLimit = ConcurrencyLimit(
-            "customer_id", max_concurrent=1
-        ),
-    ):
-        customers_executed.add(customer_id)
-        await asyncio.sleep(0.01)
-
-    # Schedule tasks for multiple customers to ensure branch coverage
-    for customer_id in [1, 2, 3]:
-        await docket.add(track_customer_execution)(customer_id=customer_id)
-
-    async with Worker(docket, concurrency=5) as worker:
-        await worker.run_until_finished()
-
-    # All customer branches should execute
-    assert customers_executed == {1, 2, 3}
-
-
-async def test_worker_concurrency_cleanup_on_success(docket: Docket):
-    """Test that concurrency slots are released when tasks complete successfully"""
-    completed_tasks: list[int] = []
-
-    async def successful_task(
-        customer_id: int,
-        concurrency: ConcurrencyLimit = ConcurrencyLimit(
-            "customer_id", max_concurrent=1
-        ),
-    ):
-        completed_tasks.append(customer_id)
-        await asyncio.sleep(0.01)
-
-    # Schedule multiple tasks that would block if slots aren't released
-    await docket.add(successful_task)(customer_id=1)
-    await docket.add(successful_task)(customer_id=1)
-    await docket.add(successful_task)(customer_id=1)
-
-    async with Worker(docket) as worker:
-        await worker.run_until_finished()
-
-    # All tasks should complete if slots are properly released
-    assert len(completed_tasks) == 3
-    assert all(customer_id == 1 for customer_id in completed_tasks)
-
-
-async def test_worker_concurrency_cleanup_on_failure(docket: Docket):
-    """Test that concurrency slots are released when tasks fail"""
-    execution_results: list[tuple[str, int, bool]] = []
-
-    async def task_that_may_fail(
-        customer_id: int,
-        should_fail: bool,
-        concurrency: ConcurrencyLimit = ConcurrencyLimit(
-            "customer_id", max_concurrent=1
-        ),
-    ):
-        execution_results.append(("executed", customer_id, should_fail))
-        await asyncio.sleep(0.01)
-
-        if should_fail:
-            raise ValueError("Intentional test failure")
-
-    # Schedule failing and successful tasks
-    await docket.add(task_that_may_fail)(customer_id=1, should_fail=True)
-    await docket.add(task_that_may_fail)(customer_id=1, should_fail=False)
-    await docket.add(task_that_may_fail)(customer_id=1, should_fail=False)
-
-    async with Worker(docket) as worker:
-        await worker.run_until_finished()
-
-    # All tasks should execute despite the failure
-    assert len(execution_results) == 3
-    failed_tasks = [r for r in execution_results if r[2] is True]
-    successful_tasks = [r for r in execution_results if r[2] is False]
-    assert len(failed_tasks) == 1
-    assert len(successful_tasks) == 2
-
-
-async def test_worker_concurrency_limits_different_scopes(docket: Docket):
-    """Test that concurrency limits work correctly with different scopes"""
-    task_executions: list[tuple[str, int]] = []
-
-    async def scoped_task(
-        customer_id: int,
-        scope_name: str,
-        concurrency: ConcurrencyLimit = ConcurrencyLimit(
-            "customer_id", max_concurrent=1, scope="custom_scope"
-        ),
-    ):
-        task_executions.append((scope_name, customer_id))
-        await asyncio.sleep(0.01)
-
-    async def default_scoped_task(
-        customer_id: int,
-        concurrency: ConcurrencyLimit = ConcurrencyLimit(
-            "customer_id", max_concurrent=1
-        ),
-    ):
-        task_executions.append(("default", customer_id))
-        await asyncio.sleep(0.01)
-
-    # Tasks with custom scope should be isolated from default scope
-    await docket.add(scoped_task)(customer_id=1, scope_name="custom")
-    await docket.add(default_scoped_task)(customer_id=1)
-
-    async with Worker(docket, concurrency=5) as worker:
-        await worker.run_until_finished()
-
-    assert len(task_executions) == 2
-    assert ("custom", 1) in task_executions
-    assert ("default", 1) in task_executions
-
-
-async def test_worker_concurrency_refresh_mechanism_integration(docket: Docket):
-    """Test that concurrency refresh mechanism works in practice"""
-    long_running_started = False
-    quick_task_completed = False
-
-    async def long_running_task(
-        customer_id: int,
-        concurrency: ConcurrencyLimit = ConcurrencyLimit(
-            "customer_id", max_concurrent=1
-        ),
-    ):
-        nonlocal long_running_started
-        long_running_started = True
-        # Long enough to trigger refresh mechanisms
-        await asyncio.sleep(0.1)
-
-    async def quick_task(
-        customer_id: int,
-        concurrency: ConcurrencyLimit = ConcurrencyLimit(
-            "customer_id", max_concurrent=1
-        ),
-    ):
-        nonlocal quick_task_completed
-        quick_task_completed = True
-
-    # Schedule a long-running task followed by a quick one
-    await docket.add(long_running_task)(customer_id=1)
-    await docket.add(quick_task)(customer_id=1)
-
-    # Test concurrency mechanism
-    worker = Worker(docket)
-
-    async with worker:
-        await worker.run_until_finished()
-
-    assert long_running_started
-    assert quick_task_completed
-
-
-async def test_worker_concurrency_with_task_failures(docket: Docket):
-    """Test that concurrency slots are properly released when tasks fail"""
-    execution_count = 0
-    failure_count = 0
-
-    async def failing_task(
-        customer_id: int,
-        should_fail: bool,
-        concurrency: ConcurrencyLimit = ConcurrencyLimit(
-            "customer_id", max_concurrent=1
-        ),
-    ):
-        nonlocal execution_count, failure_count
-        execution_count += 1
-        await asyncio.sleep(0.01)
-
-        if should_fail:
-            failure_count += 1
-            raise ValueError("Task failed intentionally")
-
-    # Schedule failing and successful tasks
-    await docket.add(failing_task)(customer_id=1, should_fail=True)
-    await docket.add(failing_task)(customer_id=1, should_fail=False)
-    await docket.add(failing_task)(customer_id=1, should_fail=False)
-
-    async with Worker(docket) as worker:
-        await worker.run_until_finished()
-
-    # All tasks should have executed despite the failure
-    assert execution_count == 3
-    assert failure_count == 1
-
-
-async def test_worker_concurrency_multiple_workers_coordination(docket: Docket):
-    """Test that multiple workers coordinate concurrency limits correctly"""
-    worker1_executions = 0
-    worker2_executions = 0
-    total_concurrent = 0
-    max_concurrent_observed = 0
-
-    async def coordinated_task(
-        customer_id: int,
-        worker_name: str,
-        concurrency: ConcurrencyLimit = ConcurrencyLimit(
-            "customer_id", max_concurrent=2
-        ),
-    ):
-        nonlocal total_concurrent, max_concurrent_observed
-        nonlocal worker1_executions, worker2_executions
-
-        total_concurrent += 1
-        max_concurrent_observed = max(max_concurrent_observed, total_concurrent)
-
-        if worker_name == "worker1":
-            worker1_executions += 1
-        else:
-            worker2_executions += 1
-
-        await asyncio.sleep(0.02)
-        total_concurrent -= 1
-
-    # Schedule many tasks for the same customer
-    for _ in range(4):
-        await docket.add(coordinated_task)(customer_id=1, worker_name="worker1")
-    for _ in range(4):
-        await docket.add(coordinated_task)(customer_id=1, worker_name="worker2")
-
-    # Run with two workers
-    worker1 = Worker(docket, name="worker1", concurrency=5)
-    worker2 = Worker(docket, name="worker2", concurrency=5)
-
-    async with worker1, worker2:
-        await asyncio.gather(worker1.run_until_finished(), worker2.run_until_finished())
-
-    # Both workers should have executed tasks
-    assert worker1_executions + worker2_executions == 8
-    # But concurrency limit should have been respected
-    assert max_concurrent_observed <= 2
-
-
-async def test_worker_concurrency_refresh_handles_redis_errors(docket: Docket):
-    """Test that concurrency refresh mechanism handles Redis errors gracefully"""
-    task_completed = False
-
-    async def task_with_concurrency(
-        customer_id: int,
-        concurrency: ConcurrencyLimit = ConcurrencyLimit(
-            "customer_id", max_concurrent=1
-        ),
-    ):
-        nonlocal task_completed
-        await asyncio.sleep(0.02)  # Long enough to trigger refresh
-        task_completed = True
-
-    await docket.add(task_with_concurrency)(customer_id=1)
-
-    # Create worker to test error handling (with short reconnection delay for testing)
-    worker = Worker(docket, reconnection_delay=timedelta(milliseconds=10))
-
-    # Mock Redis to occasionally fail during refresh operations
-    error_count = 0
-    original_redis = docket.redis
-
-    @asynccontextmanager
-    async def flaky_redis():
-        nonlocal error_count
-        if error_count == 1:  # Fail once during refresh
-            error_count += 1
-            raise ConnectionError("Simulated Redis error")
-        error_count += 1
-        async with original_redis() as redis:
-            yield redis
-
-    # Test should complete despite Redis errors in refresh mechanism
-    with patch.object(docket, "redis", flaky_redis):
-        async with worker:
-            await worker.run_until_finished()
-
-    assert task_completed
-
-
-async def test_worker_concurrency_with_quick_tasks(docket: Docket):
-    """Test that quick tasks complete without triggering complex cleanup paths"""
-    completed_tasks = 0
-
-    async def quick_task(
-        customer_id: int,
-        concurrency: ConcurrencyLimit = ConcurrencyLimit(
-            "customer_id", max_concurrent=2
-        ),
-    ):
-        nonlocal completed_tasks
-        # Very quick task to test normal execution path
-        completed_tasks += 1
-
-    # Schedule multiple quick tasks
-    for _ in range(5):
-        await docket.add(quick_task)(customer_id=1)
-
-    async with Worker(docket) as worker:
-        await worker.run_until_finished()
-
-    assert completed_tasks == 5
-
-
-async def test_worker_handles_concurrent_task_cleanup_gracefully(docket: Docket):
-    """Test that worker handles task cleanup correctly under concurrent execution"""
-    cleanup_success = True
-    task_count = 0
-
-    async def cleanup_test_task(
-        customer_id: int,
-        should_fail: bool = False,
-        concurrency: ConcurrencyLimit = ConcurrencyLimit(
-            "customer_id", max_concurrent=1
-        ),
-    ):
-        nonlocal task_count, cleanup_success
-        task_count += 1
-        try:
-            await asyncio.sleep(0.01)
-            if should_fail:
-                raise ValueError("Test exception for coverage")
-        except Exception:
-            cleanup_success = False
-            raise
-
-    # Schedule tasks that will test cleanup paths - some will fail
-    for _ in range(2):
-        await docket.add(cleanup_test_task)(customer_id=1, should_fail=False)
-
-    # Add one task that will fail to trigger exception handling
-    await docket.add(cleanup_test_task)(customer_id=1, should_fail=True)
-
-    async with Worker(docket) as worker:
-        await worker.run_until_finished()
-
-    # Should have executed all tasks (2 successful, 1 failed)
-    assert task_count == 3
-    # cleanup_success will be False because one task failed, which is expected
-    assert not cleanup_success
-
-
-async def test_worker_concurrency_with_dependencies_integration(docket: Docket):
-    """Test that concurrency limits work correctly with dependency injection"""
-    task_completed = False
-    current_worker_name = None
-
-    async def task_with_dependencies(
-        customer_id: int,
-        worker: Worker = CurrentWorker(),
-        concurrency: ConcurrencyLimit = ConcurrencyLimit(
-            "customer_id", max_concurrent=1
-        ),
-    ):
-        nonlocal task_completed, current_worker_name
-        current_worker_name = worker.name
-        await asyncio.sleep(0.01)
-        task_completed = True
-
-    await docket.add(task_with_dependencies)(customer_id=1)
-
-    async with Worker(docket, name="test-worker") as worker:
-        await worker.run_until_finished()
-
-    assert task_completed
-    assert current_worker_name == "test-worker"
-
-
-async def test_worker_concurrency_robustness_under_stress(docket: Docket):
-    """Test that concurrency management remains robust under stress conditions"""
-    successful_executions = 0
-    max_concurrent = 0
-    current_concurrent = 0
-
-    async def stress_task(
-        customer_id: int,
-        concurrency: ConcurrencyLimit = ConcurrencyLimit(
-            "customer_id", max_concurrent=3
-        ),
-    ):
-        nonlocal successful_executions, max_concurrent, current_concurrent
-        current_concurrent += 1
-        max_concurrent = max(max_concurrent, current_concurrent)
-
-        try:
-            await asyncio.sleep(0.005)  # Brief work simulation
-            successful_executions += 1
-        finally:
-            current_concurrent -= 1
-
-    # Schedule many tasks quickly
-    for _ in range(20):
-        await docket.add(stress_task)(customer_id=1)
-
-    # Process with multiple workers
-    worker1 = Worker(docket, name="worker1", concurrency=10)
-    worker2 = Worker(docket, name="worker2", concurrency=10)
-
-    # Use multiple workers to stress-test the concurrency mechanism
-
-    async with worker1, worker2:
-        await asyncio.gather(worker1.run_until_finished(), worker2.run_until_finished())
-
-    assert successful_executions == 20
-    assert max_concurrent <= 3  # Concurrency limit respected
-
-
-async def test_worker_graceful_shutdown_with_concurrency_management(docket: Docket):
-    """Test that workers shut down gracefully while managing concurrency"""
-    task_started = False
-    shutdown_completed = False
-
-    async def simple_task():
-        nonlocal task_started
-        task_started = True
-        # Quick task - we just need to verify shutdown works
-        await asyncio.sleep(0.01)
-
-    # Schedule a simple task
-    await docket.add(simple_task)()
-
-    async with Worker(docket) as worker:
-        # Start worker in background
-        asyncio.create_task(worker.run_until_finished())
-
-        # Give it a moment to process the task
-        await asyncio.sleep(0.05)
-
-        # The key test: worker should shut down gracefully when context exits
-        # The async context manager exit triggers shutdown
-
-    # If we get here, shutdown completed successfully
-    shutdown_completed = True
-
-    # Test passed if we reached this point without exceptions
-    assert shutdown_completed, "Worker should shut down gracefully"
-
-
-async def test_worker_concurrency_error_handling_during_execution(docket: Docket):
-    """Test that concurrency management handles errors gracefully during task execution"""
-    tasks_executed = 0
-    error_count = 0
-
-    async def task_that_may_error(
-        customer_id: int,
-        should_error: bool,
-        concurrency: ConcurrencyLimit = ConcurrencyLimit(
-            "customer_id", max_concurrent=1
-        ),
-    ):
-        nonlocal tasks_executed, error_count
-        tasks_executed += 1
-
-        if should_error:
-            error_count += 1
-            raise RuntimeError("Task execution error")
-
-    # Schedule both failing and successful tasks
-    await docket.add(task_that_may_error)(customer_id=1, should_error=True)
-    await docket.add(task_that_may_error)(customer_id=1, should_error=False)
-
-    # Worker should handle errors gracefully and continue processing
-    async with Worker(docket) as worker:
-        await worker.run_until_finished()
-
-    # Both tasks should have executed (one failed, one succeeded)
-    assert tasks_executed == 2
-    assert error_count == 1
-
-
-async def test_worker_concurrency_cleanup_after_task_completion(docket: Docket):
-    """Test that concurrency slots are properly cleaned up after task completion"""
-    cleanup_verified = False
-
-    async def task_with_cleanup_verification(
-        customer_id: int,
-        concurrency: ConcurrencyLimit = ConcurrencyLimit(
-            "customer_id", max_concurrent=1
-        ),
-    ):
-        # Simulate task work
-        await asyncio.sleep(0.01)
-
-    # Schedule multiple tasks to test cleanup
-    await docket.add(task_with_cleanup_verification)(customer_id=1)
-    await docket.add(task_with_cleanup_verification)(customer_id=1)
-
-    async with Worker(docket) as worker:
-        await worker.run_until_finished()
-        # After completion, verify Redis state is clean
-        async with docket.redis() as redis:
-            # Check that concurrency tracking keys are cleaned up
-            await redis.keys(f"{docket.name}:concurrency:*")  # type: ignore
-            # Keys may exist but should not have stale entries
-            cleanup_verified = True
-
-    assert cleanup_verified
-
-
-async def test_worker_concurrency_edge_cases(docket: Docket):
-    """Test edge cases in concurrency management"""
-    edge_case_handled = True
-
-    async def edge_case_task(
-        customer_id: int,
-        concurrency: ConcurrencyLimit = ConcurrencyLimit(
-            "customer_id", max_concurrent=1
-        ),
-    ):
-        # Test edge case: very quick task completion
-        pass
-
-    # Rapid scheduling and execution
-    for _ in range(5):
-        await docket.add(edge_case_task)(customer_id=1)
-
-    async with Worker(docket) as worker:
-        await worker.run_until_finished()
-
-    assert edge_case_handled
-
-
 async def test_worker_timeout_exceeds_redelivery_timeout(docket: Docket):
     """Test worker handles user timeout longer than redelivery timeout."""
 
@@ -1330,27 +561,6 @@ async def test_worker_concurrency_no_limit_with_custom_docket(docket: Docket):
     assert task_executed
 
 
-async def test_worker_concurrency_missing_argument_early_return(docket: Docket):
-    """Test early return when concurrency argument is missing."""
-    task_executed = False
-
-    async def task_missing_concurrency_arg(
-        concurrency: ConcurrencyLimit = ConcurrencyLimit(
-            "missing_param", max_concurrent=1
-        ),
-    ):
-        nonlocal task_executed
-        task_executed = True
-
-    # Call task without the required parameter - this should trigger the missing argument path
-    await docket.add(task_missing_concurrency_arg)()
-
-    async with Worker(docket) as worker:
-        await worker.run_until_finished()
-
-    assert task_executed
-
-
 async def test_worker_no_concurrency_dependency_in_function(docket: Docket):
     """Test _can_start_task with function that has no concurrency dependency."""
 
@@ -1375,93 +585,6 @@ async def test_worker_no_concurrency_dependency_in_function(docket: Docket):
             # This should return True immediately
             result = await worker._can_start_task(redis, execution)  # type: ignore[reportPrivateUsage]
             assert result is True
-
-
-async def test_worker_no_concurrency_dependency_in_release(docket: Docket):
-    """Test _release_concurrency_slot with function that has no concurrency dependency."""
-
-    async def task_without_concurrency_dependency():
-        await asyncio.sleep(0.001)
-
-    await task_without_concurrency_dependency()
-
-    async with Worker(docket) as worker:
-        # Create execution for task without concurrency dependency
-        execution = Execution(
-            docket=docket,
-            function=task_without_concurrency_dependency,
-            args=(),
-            kwargs={},
-            when=datetime.now(timezone.utc),
-            key="test_key",
-            attempt=1,
-        )
-
-        async with docket.redis() as redis:
-            # This should return immediately (line 902)
-            await worker._release_concurrency_slot(redis, execution)  # type: ignore[reportPrivateUsage]
-
-
-async def test_worker_missing_concurrency_argument_in_release(docket: Docket):
-    """Test _release_concurrency_slot when concurrency argument is missing."""
-
-    async def task_with_missing_arg(
-        concurrency: ConcurrencyLimit = ConcurrencyLimit(
-            "nonexistent_param", max_concurrent=1
-        ),
-    ):
-        await asyncio.sleep(0.001)
-
-    await task_with_missing_arg()
-
-    async with Worker(docket) as worker:
-        # Create execution that doesn't have the required parameter
-        execution = Execution(
-            docket=docket,
-            function=task_with_missing_arg,
-            args=(),
-            kwargs={},  # Missing the required parameter
-            when=datetime.now(timezone.utc),
-            key="test_key",
-            attempt=1,
-        )
-
-        async with docket.redis() as redis:
-            # This should return immediately due to KeyError (lines 907-908)
-            await worker._release_concurrency_slot(redis, execution)  # type: ignore[reportPrivateUsage]
-
-
-async def test_worker_concurrency_missing_argument_in_can_start(docket: Docket):
-    """Test _can_start_task with missing concurrency argument during execution."""
-
-    async def task_with_missing_concurrency_arg(
-        concurrency: ConcurrencyLimit = ConcurrencyLimit(
-            "missing_param", max_concurrent=1
-        ),
-    ):
-        await asyncio.sleep(0.001)
-
-    await task_with_missing_concurrency_arg()
-
-    # Register the task
-    docket.register(task_with_missing_concurrency_arg)
-
-    async with Worker(docket) as worker:
-        # Create execution without the required parameter
-        execution = Execution(
-            docket=docket,
-            function=task_with_missing_concurrency_arg,
-            args=(),
-            kwargs={},  # Missing the required "missing_param"
-            when=datetime.now(timezone.utc),
-            key="test_key",
-            attempt=1,
-        )
-
-        async with docket.redis() as redis:
-            # This should hit the KeyError path in lines 842-844
-            result = await worker._can_start_task(redis, execution)  # type: ignore[reportPrivateUsage]
-            assert result is True  # Should return True (let task fail naturally)
 
 
 async def test_worker_exception_before_dependencies(docket: Docket):
@@ -1500,31 +623,6 @@ async def test_worker_exception_before_dependencies(docket: Docket):
 
     # The task function shouldn't run via worker due to dependency failure
     assert task_failed is False
-
-
-async def test_finally_block_releases_concurrency_on_success(docket: Docket):
-    """Test that concurrency slot is released when task completes successfully."""
-    task_completed = False
-
-    async def successful_task(
-        customer_id: int,
-        concurrency: ConcurrencyLimit = ConcurrencyLimit(
-            "customer_id", max_concurrent=1
-        ),
-    ):
-        nonlocal task_completed
-        await asyncio.sleep(0.01)
-        task_completed = True
-
-    # Schedule two tasks that would block each other if slots aren't released
-    await docket.add(successful_task)(customer_id=1)
-    await docket.add(successful_task)(customer_id=1)
-
-    async with Worker(docket) as worker:
-        await worker.run_until_finished()
-
-    # If both tasks completed, the finally block successfully released slots
-    assert task_completed
 
 
 async def test_replacement_race_condition_stream_tasks(
